Read GDAL raster data as RGB triples or hex colours

gdalio_data_hex and gdalio_data_rgb are a little strange in that they return a vector of character strings and a list of numeric values respectively.

gdalio_data_rgb(dsn, bands = 1:3, ...)

gdalio_data_hex(dsn, bands = 1:3, ...)

Arguments

dsn

character string, raster source understood by GDAL

bands

bands to read, assumes 1:3 (can be 1:4 or any ordering)

...

arguments passed to 'vapour::vapour_warp_raster'

Value

'gdalio_data_rgb()' a list of integer vectors, 'gdalio_data_hex()' a character vector of "#" colours